+++ include/slg_msgs/polygon.hpp
@@ -1,7 +1,7 @@
 /*
  * POLYGON CLASS
  *
- * Copyright (c) 2020-2022 Alberto José Tudela Roldán <ajtudela@gmail.com>
+ * Copyright (c) 2020-2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of slg_msgs.
  * 
@@ -55,6 +55,10 @@ struct Edge{
 		Point2D difPA = a - p;
 		return fabs(difAB.x * difPA.y - difAB.y * difPA.x) / difAB.length();
 	}
+
+	bool self() const{
+		return a == b;
+	}
 };
 
 class Polygon{
@@ -105,6 +109,7 @@ class Polygon{
 		std::vector<Edge> get_edges() 	const { return edges; }
 		Edge get_edge(int e) 			const{ return edges[e]; }
 		void add_edge(Edge edge) 			{ edges.push_back(edge); }
+		void add_edge(Point2D a, Point2D b) { edges.push_back({a, b}); }
 
 		bool contains(const Point2D & p) const{
 			auto c = 0;
@@ -119,6 +124,45 @@ class Polygon{
 			}
 			Point2D sum = std::accumulate(points.begin(), points.end(), Point2D(0.0, 0.0));
 			return sum / points.size();
+		}
+
+		bool is_closed() const{
+			return edges.front().a == edges.back().b;
+		}
+
+		void close(){
+			if (!is_closed()){
+				edges.push_back({edges.back().b, edges.front().a});
+			}
+		}
+
+		void add_point(const Point2D & p){
+			// If the polygon is empty, add the first point
+			if (edges.empty()){
+				edges.push_back({p, p});
+			}else{
+				// If the last edge is a self edge, add a new point to it
+				if (edges.back().self()){
+					edges.back().b = p;
+				}else{
+					// If the polygon is closed, remove the last edge
+					if (is_closed()){
+						edges.pop_back();
+					}
+					// Add the new point
+					edges.push_back({edges.back().b, p});
+					// Close the polygon
+					close();
+				}
+			}
+		}
+
+		std::vector<Point2D> get_points() const{
+			std::vector<Point2D> points;
+			for (const auto & edge : edges){
+				points.push_back(edge.a);
+			}
+			return points;
 		}
 
 	private: